Camera: use multiple buffers in reading from V4L2 device
We had been using a single buffer to read from the V4L2 device,
which was causing the Camera to only get half the available frames
(since the kernel can't write to a buffer while it's in use by the
application).

Not such a big deal with the device is 60fps, but a bigger impact
with a 10fps YUYV webcam.